+++ lib/src/controls/form/auto_suggest_box.dart
@@ -15,6 +15,11 @@ typedef OnChangeAutoSuggestBox<T> = void Function(
   TextChangedReason reason,
 );
 
+typedef AutoSuggestBoxItemBuilder<T> = Widget Function(
+  BuildContext context,
+  AutoSuggestBoxItem<T> item,
+);
+
 enum TextChangedReason {
   /// Whether the text in an [AutoSuggestBox] was changed by user input
   userInput,
@@ -53,6 +58,11 @@ class AutoSuggestBoxItem<T> {
   /// Called when this item is selected
   final VoidCallback? onSelected;
 
+  /// {@macro fluent_ui.controls.inputs.HoverButton.semanticLabel}
+  ///
+  /// If not provided, [label] is used
+  final String? semanticLabel;
+
   bool _selected = false;
 
   /// Creates an auto suggest box item
@@ -62,6 +72,7 @@ class AutoSuggestBoxItem<T> {
     this.child,
     this.onFocusChange,
     this.onSelected,
+    this.semanticLabel,
   });
 
   @override
@@ -96,6 +107,7 @@ class AutoSuggestBox<T> extends StatefulWidget {
     this.controller,
     this.onChanged,
     this.onSelected,
+    this.itemBuilder,
     this.noResultsFoundBuilder,
     this.sorter,
     this.leadingIcon,
@@ -134,6 +146,7 @@ class AutoSuggestBox<T> extends StatefulWidget {
     this.controller,
     this.onChanged,
     this.onSelected,
+    this.itemBuilder,
     this.noResultsFoundBuilder,
     this.sorter,
     this.leadingIcon,
@@ -177,6 +190,11 @@ class AutoSuggestBox<T> extends StatefulWidget {
 
   /// Called when the user selected a value.
   final ValueChanged<AutoSuggestBoxItem<T>>? onSelected;
+
+  /// A callback function that builds the items in the overlay.
+  ///
+  /// Use [noResultsFoundBuilder] to build the overlay when no item is provided
+  final AutoSuggestBoxItemBuilder? itemBuilder;
 
   /// Widget to be displayed when none of the items fit the [sorter]
   final WidgetBuilder? noResultsFoundBuilder;
@@ -518,6 +536,7 @@ class _AutoSuggestBoxState<T> extends State<AutoSuggestBox<T>> {
                 node: overlayNode,
                 controller: controller,
                 items: widget.items,
+                itemBuilder: widget.itemBuilder,
                 focusStream: _focusStreamController.stream,
                 itemsStream: _dynamicItemsController.stream,
                 sorter: sorter,
@@ -745,6 +764,7 @@ class _AutoSuggestBoxOverlay<T> extends StatefulWidget {
   const _AutoSuggestBoxOverlay({
     super.key,
     required this.items,
+    required this.itemBuilder,
     required this.controller,
     required this.onSelected,
     required this.node,
@@ -756,6 +776,7 @@ class _AutoSuggestBoxOverlay<T> extends StatefulWidget {
   });
 
   final List<AutoSuggestBoxItem<T>> items;
+  final AutoSuggestBoxItemBuilder<T>? itemBuilder;
   final TextEditingController controller;
   final ValueChanged<AutoSuggestBoxItem<T>> onSelected;
   final FocusScopeNode node;
@@ -867,11 +888,13 @@ class _AutoSuggestBoxOverlayState<T> extends State<_AutoSuggestBoxOverlay<T>> {
                     itemCount: sortedItems.length,
                     itemBuilder: (context, index) {
                       final item = sortedItems[index];
-                      return _AutoSuggestBoxOverlayTile(
-                        text: item.child ?? Text(item.label),
-                        selected: item._selected || widget.node.hasFocus,
-                        onSelected: () => widget.onSelected(item),
-                      );
+                      return widget.itemBuilder?.call(context, item) ??
+                          _AutoSuggestBoxOverlayTile(
+                            text: item.child ?? Text(item.label),
+                            semanticLabel: item.semanticLabel ?? item.label,
+                            selected: item._selected || widget.node.hasFocus,
+                            onSelected: () => widget.onSelected(item),
+                          );
                     },
                   );
                 }
@@ -890,11 +913,13 @@ class _AutoSuggestBoxOverlayTile extends StatefulWidget {
     required this.text,
     this.selected = false,
     this.onSelected,
+    this.semanticLabel,
   });
 
   final Widget text;
   final VoidCallback? onSelected;
   final bool selected;
+  final String? semanticLabel;
 
   @override
   State<_AutoSuggestBoxOverlayTile> createState() =>
@@ -926,6 +951,7 @@ class __AutoSuggestBoxOverlayTileState extends State<_AutoSuggestBoxOverlayTile>
     final theme = FluentTheme.of(context);
 
     return ListTile.selectable(
+      semanticLabel: widget.semanticLabel,
       title: EntrancePageTransition(
         animation: Tween<double>(
           begin: 0.75,